MM-12901 Create showSearchModal action and use for hashtags
