+++ app/components/markdown/hashtag/hashtag.test.js
@@ -11,8 +11,13 @@ describe('Hashtag', () => {
     const baseProps = {
         hashtag: 'test',
         linkStyle: {color: 'red'},
-        navigator: {},
-        theme: {},
+        navigator: {
+            dismissAllModals: jest.fn(),
+            popToRoot: jest.fn(),
+        },
+        actions: {
+            showSearchModal: jest.fn(),
+        },
     };
 
     test('should match snapshot', () => {
@@ -24,11 +29,6 @@ describe('Hashtag', () => {
     test('should open hashtag search on click', () => {
         const props = {
             ...baseProps,
-            navigator: {
-                dismissAllModals: jest.fn(),
-                popToRoot: jest.fn(),
-                showModal: jest.fn(),
-            },
         };
 
         const wrapper = shallow(<Hashtag {...props}/>);
@@ -37,22 +37,12 @@ describe('Hashtag', () => {
 
         expect(props.navigator.dismissAllModals).toHaveBeenCalled();
         expect(props.navigator.popToRoot).toHaveBeenCalled();
-        expect(props.navigator.showModal).toHaveBeenCalledWith(expect.objectContaining({
-            screen: 'Search',
-            passProps: {
-                initialValue: '#test',
-            },
-        }));
+        expect(props.actions.showSearchModal).toHaveBeenCalledWith(props.navigator, '#test');
     });
 
     test('should call onHashtagPress if provided', () => {
         const props = {
             ...baseProps,
-            navigator: {
-                dismissAllModals: jest.fn(),
-                popToRoot: jest.fn(),
-                showModal: jest.fn(),
-            },
             onHashtagPress: jest.fn(),
         };
 
@@ -62,7 +52,7 @@ describe('Hashtag', () => {
 
         expect(props.navigator.dismissAllModals).not.toBeCalled();
         expect(props.navigator.popToRoot).not.toBeCalled();
-        expect(props.navigator.showModal).not.toBeCalled();
+        expect(props.actions.showSearchModal).not.toBeCalled();
 
         expect(props.onHashtagPress).toBeCalled();
     });
