+++ app/components/markdown/markdown.js
@@ -232,7 +232,6 @@ export default class Markdown extends PureComponent {
                 linkStyle={this.props.textStyles.link}
                 onHashtagPress={this.props.onHashtagPress}
                 navigator={this.props.navigator}
-                theme={this.props.theme}
             />
         );
     }
